refactored linear submodule to policies and value_functions
