+++ doc/policies/updateable.rst
@@ -4,9 +4,9 @@ Updateable Policies
 .. autosummary::
     :nosignatures:
 
-    keras_gym.policies.SoftmaxPolicy
+    keras_gym.policies.GenericSoftmaxPolicy
     keras_gym.policies.LinearSoftmaxPolicy
 
 
-.. autoclass:: keras_gym.policies.SoftmaxPolicy
+.. autoclass:: keras_gym.policies.GenericSoftmaxPolicy
 .. autoclass:: keras_gym.policies.LinearSoftmaxPolicy
