--- keras_gym/base/function_approximators/__init__.py
@@ -1,3 +1,0 @@
-# flake8: noqa
-from .base import *
-from .linear import *
