+++ keras_gym/policies/__init__.py
@@ -1,3 +1,4 @@
 # flake8: noqa
-from ..base.function_approximators import SoftmaxPolicy, LinearSoftmaxPolicy
+from ..base.function_approximators.generic import GenericSoftmaxPolicy
 from .value_based import *
+from .linear import *
